+++ R/RSTheme_light.R
@@ -1,0 +1,9 @@
+#' @title set a light theme
+#' @description only the dark theme is currently under development. This
+#' function is a placeholder for a future light version of R.AlphA themes.
+#'
+#' @return nothing. Applies chrome theme
+#' @export
+
+
+RSTheme_light <- function() rstudioapi::applyTheme("chrome")